[cm] Refactoring, scattering 1d done
