+++ python/jtfst.py
@@ -1,85 +1,15 @@
 import logging
 import os
-from typing import Optional, List
 
 import torch as tr
-import torch.nn.functional as F
 import torchaudio
-from torch import Tensor as T
-from tqdm import tqdm
 
-from scalogram import plot_scalogram, calc_scalogram_fd, calc_scalogram_td
+from scalogram_1d import plot_scalogram_1d, calc_scalogram_1d
 from wavelets import MorletWavelet, make_wavelet_bank
 
 logging.basicConfig()
 log = logging.getLogger(__name__)
 log.setLevel(level=os.environ.get("LOGLEVEL", "INFO"))
-
-
-def calc_jtfst_td(scalogram: T, wavelet_bank: List[T], take_modulus: bool = True) -> T:
-    assert scalogram.ndim == 3
-    assert wavelet_bank
-
-    scalogram = scalogram.unsqueeze(1)  # Image with 1 channel
-    scalogram_complex = scalogram.to(mw.dtype)
-    convs = []
-    for wavelet in tqdm(wavelet_bank):
-        assert wavelet.ndim == 2
-        kernel = wavelet.view(1, 1, *wavelet.shape)
-        out = F.conv2d(scalogram_complex, kernel, stride=(1, 1), padding="same")
-        convs.append(out)
-
-    jtfst = tr.cat(convs, dim=1)
-    if take_modulus:
-        jtfst = tr.abs(jtfst)
-
-    return jtfst
-
-
-def calc_jtfst_fd(scalogram: T,
-                  wavelet_bank: List[T],
-                  max_f_dim: Optional[int] = None,
-                  max_t_dim: Optional[int] = None,
-                  take_modulus: bool = True) -> T:
-    assert scalogram.ndim == 3
-    if max_f_dim is None:
-        max_f_dim = max([w.size(0) for w in wavelet_bank])
-    if max_t_dim is None:
-        max_t_dim = max([w.size(1) for w in wavelet_bank])
-    max_f_padding = max_f_dim // 2
-    max_t_padding = max_t_dim // 2
-    # TODO(cm): check why we can get away with only padding the front
-    scalogram = F.pad(scalogram, (max_t_padding, 0, max_f_padding, 0))
-    log.debug("scalogram fft")
-    scalogram_fd = tr.fft.fft2(scalogram).unsqueeze(1)
-
-    log.debug("making kernels")
-    kernels = []
-    for wavelet in wavelet_bank:
-        assert wavelet.ndim == 2
-        top_padding = max_f_padding - wavelet.size(-2) // 2
-        bottom_padding = scalogram_fd.size(-2) - wavelet.size(-2) - top_padding
-        left_padding = max_t_padding - wavelet.size(-1) // 2
-        right_padding = scalogram_fd.size(-1) - wavelet.size(-1) - left_padding
-        kernel = wavelet.view(1, 1, *wavelet.shape)
-        kernel = F.pad(kernel, (left_padding, right_padding, top_padding, bottom_padding))
-        kernels.append(kernel)
-
-    kernels = tr.cat(kernels, dim=1)
-    log.debug("kernel fft")
-    kernels_fd = tr.fft.fft2(kernels)
-    log.debug("matmult")
-    kernels_fd.imag *= -1  # PyTorch does cross-correlation instead of convolution
-    out_fd = kernels_fd * scalogram_fd
-    log.debug("jtfst ifft")
-    jtfst = tr.fft.ifft2(out_fd)
-    # TODO(cm): check why removing padding from the end works empirically after IFFT
-    jtfst = jtfst[:, :, :-max_f_padding, :-max_t_padding]
-
-    if take_modulus:
-        jtfst = tr.abs(jtfst)
-
-    return jtfst
 
 
 if __name__ == "__main__":
@@ -144,12 +74,12 @@ if __name__ == "__main__":
     wavelet_bank, freqs = make_wavelet_bank(mw, J_1, Q_1, highest_freq_t=highest_freq)
 
     log.info(f"in audio.shape = {audio.shape}")
-    # scalogram = calc_scalogram_td(audio, wavelet_bank, take_modulus=True)
-    scalogram = calc_scalogram_fd(audio, wavelet_bank, take_modulus=True)
+    # scalogram = calc_scalogram_1d_td(audio, wavelet_bank, take_modulus=True)
+    scalogram = calc_scalogram_1d(audio, wavelet_bank, take_modulus=True)
     log.info(f"scalogram shape = {scalogram.shape}")
     log.info(f"scalogram mean = {tr.mean(scalogram)}")
     log.info(f"scalogram std = {tr.std(scalogram)}")
     log.info(f"scalogram max = {tr.max(scalogram)}")
     log.info(f"scalogram min = {tr.min(scalogram)}")
-    plot_scalogram(scalogram[0], title="chirp", dt=mw.dt, freqs=freqs)
-    # plot_scalogram(scalogram[1], title="flute", dt=mw.dt, freqs=freqs)
+    plot_scalogram_1d(scalogram[0], title="chirp", dt=mw.dt, freqs=freqs)
+    # plot_scalogram_1d(scalogram[1], title="flute", dt=mw.dt, freqs=freqs)
