--- python/svm_classifier.py
@@ -1,10 +1,0 @@
-import logging
-import os
-
-logging.basicConfig()
-log = logging.getLogger(__name__)
-log.setLevel(level=os.environ.get("LOGLEVEL", "INFO"))
-
-
-if __name__ == "__main__":
-    log.info("SVM classifier")
